backend/pyproject.toml
[tool.poetry]
name = "intervista-assistant-backend"
version = "0.1.0"
description = "Backend per l'assistente di interviste in tempo reale"
authors = ["Carlo Giuseppe Sergi <c.sergi@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.64.0"
python-dotenv = "^1.0.1"
pyaudio = "^0.2.14"
websocket-client = "^1.8.0"
numpy = "^2.2.3"
flask = "^3.1.0"
flask-cors = "^5.0.1"
flask-socketio = "^5.5.1"
python-socketio = "^5.12.1"
eventlet = "^0.30.0"

[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "backend.api_launcher:main" 